requirements.txt
flask
flask-cors
openai-whisper
torch
numpy
opencv-python
supabase
python-dotenv

